src/analyst/requirements.txt
flask
gunicorn
google-cloud-aiplatform
google-cloud-firestore